Refactor Stop Details Page
