+++ frontend/app/[locale]/(website)/stops/[stop_id]/page.tsx
@@ -1,6 +1,6 @@
 /* * */
 
-import LinesSingle from '@/components/stops/Single';
+import StopsDetail from '@/components/stops/StopDetail';
 import { StopsListContextProvider } from '@/contexts/StopsList.context';
 import { StopsSingleContextProvider } from '@/contexts/StopsSingle.context';
 
@@ -10,7 +10,7 @@ export default function Page({ params: { stop_id } }) {
 	return (
 		<StopsListContextProvider>
 			<StopsSingleContextProvider stopId={stop_id}>
-				<LinesSingle />
+				<StopsDetail />
 			</StopsSingleContextProvider>
 		</StopsListContextProvider>
 	);
