--- frontend/components/stops/Single/styles.module.css
@@ -1,168 +1,0 @@
-/* * */
-/* HEADING SECTION */
-
-.headingSection {
-	display: flex;
-	flex-direction: column;
-	justify-content: flex;
-    align-items: flex-start;
-	gap: var(--size-spacing-15);
-}
-
-/* * */
-/* HEADING SECTION / ROW */
-
-.headingSectionRow {
-	display: flex;
-	flex-direction: row;
-	justify-content: flex-start;
-	align-items: center;
-	gap: var(--size-spacing-15);
-}
-
-.headingTitle{
-    font-size: 20px;
-    font-weight: 700;
-    align-items: center;
-}
-
-.volumeIcon{
-    position:relative;
-    top: 6px;
-    margin-left: 8px;
-    margin-right: 8px;
-    color: var(--color-system-text-400);
-    cursor:pointer;
-}
-
-.volumeIcon:hover{
-    color: var(--color-system-text-300);
-}
-
-.favoriteIcon {
-    position:relative;
-    top: 6px;
-}
-
-.stopLocation{
-    color: var(--color-system-text-300);
-    font-size: 14px;
-    font-weight: 600;
-}
-
-.badges{
-    display: flex;
-    gap: 8px;
-}
-
-
-
-/* * */
-/* * */
-/* * */
-/* * */
-/* * */
-
-.pageWrapper{
-    display: flex;
-    flex-direction: column;
-    gap: 30px;
-    padding-top: 25px;
-}
-
-.header,.controls{
-    padding: 0px 20px 0px 20px;
-    display: flex;
-    flex-direction: column;
-    gap:10px;
-}
-
-.controls{
-    gap: 30px
-}
-
-.header{
-    padding-top: 0;
-}
-
-.header h3 {
-    font-size: 20px;
-    font-weight: 700;
-}
-
-.tag{
-    border-radius: 1000px;
-    padding: 3px 15px;
-    font-size: 20px;
-    font-weight: 700;
-}
-.lineHeader{
-    display: flex;
-    align-items: center;
-    gap: 15px;
-}
-
-.heart{
-    --fill: #0000;
-    --stroke: var(--color-system-text-400);
-}
-@media only screen and (max-width: 400px) {
-    .heart:hover{
-        --fill: #FEE2E2;
-        --stroke: #FECACA;
-    }
-}
-
-.routeSelector{
-    display: flex;
-    gap: 10px;
-    flex-direction: column;
-}
-
-.routeExplainer{
-    align-self: flex-end;
-    font-size: 10px;
-    color: var(--color-system-text-300);
-    font-style: italic;
-    display: flex;
-    align-items: center;
-}
-
-.mapAttribution{
-    display: flex;
-    gap: 4px;
-    padding-left: 8px;
-    font-size: 10px;
-    color: var(--color-system-text-300);
-}
-.mapAttribution a:hover{
-    text-decoration: underline;
-}
-
-.divisor{
-    height: 1px;
-    width: 100%;
-    background-color: var(--color-system-border-100);
-    margin: 0px 0px;
-}
-
-.input::placeholder{
-    color: var(--color-system-text-300)
-}
-.input{
-    color: var(--color-system-text-200)
-}
-
-.metrics{
-    padding-left: 20px;
-    padding-right: 20px;
-}
-.metrics h1 {
-    font-weight: 700;
-    font-size: 60px;
-}
-.metrics h3 {
-    font-size: 18px;
-    font-weight: 700;
-    color: var(--color-system-text-100)
-}
