+++ frontend/components/stops/StopDetail/styles.module.css
@@ -1,0 +1,18 @@
+/* * */
+/* HEADING SECTION */
+
+.headingSection {
+	display: flex;
+	flex-direction: column;
+	justify-content: flex-start;
+	align-items: flex-start;
+	gap: var(--size-spacing-15);
+}
+
+.headingSectionRow {
+	display: flex;
+	flex-direction: row;
+	justify-content: flex-start;
+	align-items: center;
+	gap: var(--size-spacing-15);
+}
