+++ frontend/components/stops/StopDetailToolbar/styles.module.css
@@ -1,0 +1,31 @@
+/* * */
+/* CONTAINER */
+
+.container {
+	display: grid;
+	grid-template-columns: auto 1fr;
+	gap: var(--size-spacing-10) var(--size-spacing-20);
+	align-items: center;
+	justify-content: center;
+}
+
+@media screen and (width <= 900px) {
+
+	.container {
+		grid-template-columns: 1fr;
+	}
+}
+
+/* * */
+/* PATTERN SELECTOR EXPLAINER WRAPPER */
+
+.patternSelectorExplainerWrapper {
+	justify-self: flex-end;
+}
+
+@media screen and (width <= 900px) {
+
+	.patternSelectorExplainerWrapper {
+		margin-top: var(--size-spacing-20);
+	}
+}
