+++ frontend/components/stops/StopMap/styles.module.css
@@ -16,7 +16,7 @@
 }
 
 .map{
-    height: 220px;
+    height: 100%;
     width: 100%;
 }
